+++ dom/src/main/java/react4j/dom/events/ClipboardEvent.java
@@ -1,11 +1,19 @@
 package react4j.dom.events;
 
 import elemental2.dom.DataTransfer;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsType;
 
 @JsType( isNative = true )
+@SuppressWarnings( "unused" )
 public class ClipboardEvent
-  extends SyntheticEvent
+  extends SyntheticEvent<elemental2.dom.ClipboardEvent>
 {
-  public DataTransfer clipboardData;
+  private DataTransfer clipboardData;
+
+  @JsOverlay
+  public final DataTransfer getClipboardData()
+  {
+    return clipboardData;
+  }
 }