+++ dom/src/main/java/react4j/dom/events/CompositionEvent.java
@@ -1,10 +1,19 @@
 package react4j.dom.events;
 
+import elemental2.dom.Event;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsType;
 
 @JsType( isNative = true )
+@SuppressWarnings( "unused" )
 public class CompositionEvent
-  extends SyntheticEvent
+  extends SyntheticEvent<Event>
 {
-  public String data;
+  private String data;
+
+  @JsOverlay
+  public final String getData()
+  {
+    return data;
+  }
 }