+++ dom/src/main/java/react4j/dom/events/FocusEvent.java
@@ -1,11 +1,19 @@
 package react4j.dom.events;
 
 import elemental2.dom.EventTarget;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsType;
 
 @JsType( isNative = true )
+@SuppressWarnings( "unused" )
 public class FocusEvent
-  extends SyntheticEvent
+  extends SyntheticEvent<elemental2.dom.FocusEvent>
 {
-  public EventTarget relatedTarget;
+  private EventTarget relatedTarget;
+
+  @JsOverlay
+  public final EventTarget getRelatedTarget()
+  {
+    return relatedTarget;
+  }
 }