+++ dom/src/main/java/react4j/dom/events/FormEvent.java
@@ -1,9 +1,11 @@
 package react4j.dom.events;
 
+import elemental2.dom.Event;
 import jsinterop.annotations.JsType;
 
 @JsType( isNative = true )
+@SuppressWarnings( "unused" )
 public class FormEvent
-  extends SyntheticEvent
+  extends SyntheticEvent<Event>
 {
 }