+++ dom/src/main/java/react4j/dom/events/KeyboardEvent.java
@@ -1,22 +1,91 @@
 package react4j.dom.events;
 
+import javax.annotation.Nonnull;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsType;
 
 @JsType( isNative = true )
+@SuppressWarnings( "unused" )
 public class KeyboardEvent
-  extends SyntheticEvent
+  extends SyntheticEvent<elemental2.dom.KeyboardEvent>
 {
-  public boolean altKey;
-  public int charCode;
-  public boolean ctrlKey;
-  public String key;
-  public int keyCode;
-  public String locale;
-  public int location;
-  public boolean metaKey;
-  public boolean repeat;
-  public boolean shiftKey;
-  public int which;
+  private boolean altKey;
+  private int charCode;
+  private boolean ctrlKey;
+  private String key;
+  private int keyCode;
+  private String locale;
+  private int location;
+  private boolean metaKey;
+  private boolean repeat;
+  private boolean shiftKey;
+  private int which;
 
-  public native boolean getModifierState( String key );
+  public native boolean getModifierState( @Nonnull String key );
+
+  @JsOverlay
+  public final boolean isAltKey()
+  {
+    return altKey;
+  }
+
+  @JsOverlay
+  public final int getCharCode()
+  {
+    return charCode;
+  }
+
+  @JsOverlay
+  public final boolean isCtrlKey()
+  {
+    return ctrlKey;
+  }
+
+  @JsOverlay
+  public final String getKey()
+  {
+    return key;
+  }
+
+  @JsOverlay
+  public final int getKeyCode()
+  {
+    return keyCode;
+  }
+
+  @JsOverlay
+  public final String getLocale()
+  {
+    return locale;
+  }
+
+  @JsOverlay
+  public final int getLocation()
+  {
+    return location;
+  }
+
+  @JsOverlay
+  public final boolean isMetaKey()
+  {
+    return metaKey;
+  }
+
+  @JsOverlay
+  public final boolean isRepeat()
+  {
+    return repeat;
+  }
+
+  @JsOverlay
+  public final boolean isShiftKey()
+  {
+    return shiftKey;
+  }
+
+  @JsOverlay
+  public final int getWhich()
+  {
+    return which;
+  }
 }