+++ dom/src/main/java/react4j/dom/events/TouchList.java
@@ -1,9 +1,0 @@
-package react4j.dom.events;
-
-import jsinterop.annotations.JsType;
-
-@JsType( isNative = true )
-public class TouchList
-{
-  //TODO
-}