+++ dom/src/main/java/react4j/dom/events/UIEvent.java
@@ -1,11 +1,25 @@
 package react4j.dom.events;
 
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsType;
 
 @JsType( isNative = true )
+@SuppressWarnings( "unused" )
 public class UIEvent
-  extends SyntheticEvent
+  extends SyntheticEvent<elemental2.dom.UIEvent>
 {
-  public int detail;
-  public AbstractView view;
+  private int detail;
+  private AbstractView view;
+
+  @JsOverlay
+  public final int getDetail()
+  {
+    return detail;
+  }
+
+  @JsOverlay
+  public final AbstractView getView()
+  {
+    return view;
+  }
 }